feat: move args parsing out of controller
